\name{stockplot-package}
\alias{stockplot-package}
\alias{stockplot}
\docType{package}
\title{
\packageTitle{stockplot}
}
\description{
\packageDescription{stockplot}
The stockplot package is intended for the sole purpose of comparing the relative returns over a user-specified investment horizon, indicated by
the date range inherent in the data that the user supplies. The combination of the two functions in the package, \code{relative} and
\code{plot.relative} allows the user to input a data set of their choice, which should containt the historical closing prices of two stocks of
their choice, and plot the relative returns from each stock as of the first date specified in their data set. The formatting in these functions 
is intended to accomodate typical formatting used by analysis in financial institutions whor egularly work with Excel and Bloomberg-style formatted
data.
}
\details{
\packageDESCRIPTION{stockplot}
\packageIndices{stockplot}
}
\author{Villegas, J.G}
\packageAuthor{stockplot}

Maintainer: \packageMaintainer{stockplot}

\references{
lol read a book
}
\examples{
~~ simple examples of the most important functions ~~
}
